requirements.txt
a2wsgi==1.8.0
anyio==4.1.0
asgiref==3.7.2
attrs==23.1.0
Babel==2.14.0
blinker==1.7.0
Bootstrap-Flask==2.3.3
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
connexion==3.0.2
exceptiongroup==1.2.0
Flask==3.0.0
Flask-Login==0.6.3
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
KalturaApiClient==19.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.13.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.32.0.post1
swagger_ui_bundle==1.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
WTForms==3.1.1
python-jose==3.3.0